+++ database/migrations/2024_04_01_031739_create_orders_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('user_id')->nullable();
             $table->unsignedBigInteger('status_id')->default(1);
-            $table->decimal('total');
+            $table->bigInteger('total');
             $table->foreign('user_id')->references('id')->on('users');
             $table->foreign('status_id')->references('id')->on('order_statuses');
             $table->timestamps();
